Simplify spinner logic
