+++ apps/mobile-wallet/src/components/layout/TransactionsFlatList.tsx
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ForwardedRef, forwardRef, useCallback, useEffect, useState } from 'react'
+import { ForwardedRef, forwardRef, useCallback, useState } from 'react'
 import { ActivityIndicator, FlatList, FlatListProps } from 'react-native'
 import { Portal } from 'react-native-portalize'
 import styled, { useTheme } from 'styled-components/native'
@@ -27,7 +27,7 @@ import BottomModal from '~/components/layout/BottomModal'
 import RefreshSpinner from '~/components/RefreshSpinner'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import TransactionModal from '~/screens/TransactionModal'
-import { syncAllAddressesTransactionsNextPage, syncLatestTransactions } from '~/store/addressesSlice'
+import { syncAllAddressesTransactionsNextPage } from '~/store/addressesSlice'
 import { DEFAULT_MARGIN, SCREEN_OVERFLOW } from '~/style/globalStyle'
 import { AddressConfirmedTransaction, AddressPendingTransaction, AddressTransaction } from '~/types/transactions'
 import { isPendingTx } from '~/utils/transactions'
@@ -65,10 +65,8 @@ const TransactionsFlatList = forwardRef(function TransactionsFlatList(
 
   const isLoading = useAppSelector((s) => s.addresses.loadingTransactionsNextPage)
   const allConfirmedTransactionsLoaded = useAppSelector((s) => s.confirmedTransactions.allLoaded)
-  const isLoadingLatestTxs = useAppSelector((s) => s.addresses.loadingLatestTransactions)
 
   const [txModalOpen, setTxModalOpen] = useState(false)
-  const [isSpinnerVisible, setIsSpinnerVisible] = useState(false)
   const [selectedTx, setSelectedTx] = useState<AddressConfirmedTransaction>()
 
   const renderConfirmedTransactionItem = ({ item, index }: TransactionItem) =>
@@ -94,17 +92,6 @@ const TransactionsFlatList = forwardRef(function TransactionsFlatList(
     dispatch(syncAllAddressesTransactionsNextPage({ minTxs: 10 }))
   }, [allConfirmedTransactionsLoaded, dispatch, isLoading])
 
-  useEffect(() => {
-    if (!isLoadingLatestTxs) setIsSpinnerVisible(false)
-  }, [isLoadingLatestTxs])
-
-  const refreshData = () => {
-    setIsSpinnerVisible(true)
-
-    if (!isLoadingLatestTxs) {
-      dispatch(syncLatestTransactions()).finally(() => setIsSpinnerVisible(false))
-    }
-  }
   return (
     <>
       <FlatList
@@ -117,7 +104,7 @@ const TransactionsFlatList = forwardRef(function TransactionsFlatList(
         keyExtractor={transactionKeyExtractor}
         onEndReached={loadNextTransactionsPage}
         style={{ overflow: SCREEN_OVERFLOW }}
-        refreshControl={<RefreshSpinner refreshing={isSpinnerVisible} onRefresh={refreshData} />}
+        refreshControl={<RefreshSpinner />}
         refreshing={pendingTransactions.length > 0}
         extraData={confirmedTransactions.length > 0 ? confirmedTransactions[0].hash : ''}
         ListHeaderComponent={
